RAG/requirements.txt
openai
faiss-cpu #if it fails, try conda install -c conda-forge faiss
chromadb
langchain
tiktoken
huggingface-hub
python-dotenv
pandas
sentence-transformers
transformers